attack_eye/requirements/base.txt
amqp==2.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.5.0.5
celery==4.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Django==4.0.3
django-cors-headers==3.13.0
dj-database-url-1.2.0
django-session-timeout==0.1.0
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.9.0
idna==3.4
kombu==4.6.11
mysqlclient==2.1.1
pipreqs==0.4.11
prompt-toolkit==3.0.36
PyMySQL==1.0.2
python-decouple==3.7
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
typing-extensions==4.5.0
urllib3==1.26.13
validators==0.20.0
vine==1.3.0
wcwidth==0.2.6
yarg==0.1.9
